Constructor for BrowserViz

This constructor function:

creates the BrowserViz object

initializes the httpuv web server

prepares that web server to additionally handle websocket traffic

loads a "browserFile" - an html/javascript/css web page to communicate with in your web browser

opens websocket communication between your R session and your browser

installs an optional "httpQueryProcessingFunction" to handle http (non-websocket) requests.

BrowserViz(portRange = 10000:10100, title = "BrowserViz", browserFile,
  quiet = TRUE, httpQueryProcessingFunction = NULL)

Arguments

portRange
The constructor looks for a free websocket port in this range. 15000:15100 by default

title
Used for the web browser window, "igvR" by default

browserFile
The full path to the bundled html, js and libraries, and css which constitute the browser app

quiet
A logical variable controlling verbosity during execution

httpQueryProcessingFunction
a function, default NULL, provides subclasses with the opportunity to execute code on the http server created here.

Value

An object of the BrowserViZ class